Initial work on porting block minimization from ll2x

+++ include/graph/DTree.h
@@ -16,7 +16,7 @@ namespace LL2W {
 
 		public:
 			Node *startNode = nullptr;
-			std::unordered_map<Node *, int> levels;
+			std::unordered_map<Node *, size_t> levels;
 			std::unordered_map<Node *, Node *> dominators;
 
 			/** Constructs a dominator tree from a graph. The dominator tree is a tree in which each node of the
